Update error response function, add user auth middleware
The SendErrorJsonResponse function now includes an HTTP status code parameter. A middleware has been added to handle user authorization checks, and a new function retrieves the user from their token. The user ID data type in the repository has also been updated.

+++ internal/users/data/users_repository.go
@@ -21,7 +21,7 @@ func UserWithEmail(email string) repository.SelectCriteria {
 	}
 }
 
-func UserWithID(id int) repository.SelectCriteria {
+func UserWithID(id int64) repository.SelectCriteria {
 	return func(query *bun.SelectQuery) *bun.SelectQuery {
 		return query.Where("id = ?", id)
 	}
